requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic[email]==2.9.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
